Collapse testing back into pluginlib package.
