+++ pluginlib/package.xml
@@ -25,6 +25,8 @@
   <depend>rcutils</depend>
   <depend>tinyxml2_vendor</depend>
 
+  <test_depend>ament_cmake_gtest</test_depend>
+
   <export>
     <build_type>ament_cmake</build_type>
   </export>
