+++ pluginlib/test/test_plugins.xml
