--- test_pluginlib/CHANGELOG.rst
@@ -1,8 +1,0 @@
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-Changelog for package test_pluginlib
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-2.2.0 (2018-11-20)
-------------------
-* Reorganized tests into fixture and consumer packages. (`#130 <https://github.com/ros/pluginlib/issues/130>`_)
-* Contributors: Michael Carroll
