--- test_pluginlib_fixture/CHANGELOG.rst
@@ -1,9 +1,0 @@
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-Changelog for package test_pluginlib_fixture
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-2.2.0 (2018-11-20)
-------------------
-* Reorganized tests into fixture and consumer packages. (`#130 <https://github.com/ros/pluginlib/issues/130>`_)
-* Contributors: Michael Carroll
-
